Refactor Onnx prediction to utils folder
